Raise the maximum upload size to 60GB
MAX_UPLOAD_MB alone could not deliver this — three blockers were
independent of the setting:

1. Uploads were buffered in full before the handler ran. Starlette spools
   each multipart part to a temp file, so a 60GB upload needed 60GB in the
   container's own filesystem PLUS 60GB in /data — and the disk guard only
   measures /data, so it couldn't see the half that fills first.
2. output.zip contained the source video, deflated: a third 60GB copy
   produced by compressing already-compressed H.264 for ~0% saving.
3. The job would have been killed anyway — Celery's ceiling was ~2h, and
   the stale-job reaper fails anything quiet for 30min, while scene
   detection and zipping sent no heartbeat at all.

Streaming upload (new POST /api/jobs/upload): raw body straight to disk in
8MiB chunks, hashed as it goes, nothing buffered. Verified against a real
uvicorn process — 1.5GB streamed at 80MB/s raised peak RSS by 784kB, and a
0.58GB video round-tripped byte-perfect (sha256 match). Oversized uploads
now 413 from Content-Length before a single byte transfers (measured: 58ms,
0 bytes sent) instead of after a full transfer. The multipart route stays
for URL jobs and small files; both share one job-creation helper.

ZIP excludes the source video (ZIP_INCLUDE_SOURCE_VIDEO=false): the archive
carries transcripts, frames, and metadata while the video stays on disk and
downloadable from /api/jobs/{id}/video. The manifest still describes the
file and marks it included_in_zip: false with the reason, and
metadata/zip_exclusions.json records it.

Long-job survival: 24h job ceiling, 6h per-step limits, and HeartbeatTicker
around scene detection and the zip write so the reaper stops failing
healthy work. Split the reaper's timeout in two — STALE_JOB_TIMEOUT_MINUTES
now applies only to running jobs, since with concurrency 1 a huge source
legitimately holds the queue for hours; QUEUED_JOB_TIMEOUT_HOURS (24h)
still catches a job whose enqueue never reached the broker.

Also fixed along the way: OSError/ENOSPC and client disconnects mid-upload
now clean up instead of stranding a partial file; ffprobe moved off the
event loop (it blocked the healthcheck) with a configurable timeout;
file_size_bytes widened to BigInteger so 60GB doesn't overflow on Postgres.

Disk accounting: required space is now upload x
UPLOAD_DISK_HEADROOM_MULTIPLIER (1.5) + MIN_FREE_DISK_MB, covering frames,
temp audio, and the ZIP — a 60GB upload needs ~92GB free and is rejected up
front with a 507 explaining the arithmetic rather than dying on ENOSPC.

Frontend sends files as a raw XHR body (progress unchanged), rejects
oversized files before transferring, and formats sizes in GB.

19 new tests (217 total passing), tsc + next build clean, and the
fresh-dataset E2E confirms the ZIP has no source/ member while transcripts,
frames, and a 'success' validation report are intact.

+++ backend/app/models.py
@@ -3,7 +3,7 @@ same schema works unchanged against Postgres later."""
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -69,7 +69,9 @@ class Job(Base):
     language: Mapped[str | None] = mapped_column(String(16), nullable=True)
     frame_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
 
-    file_size_bytes: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # BigInteger: a 60GB upload is ~6.4e10 bytes, well past a 32-bit INTEGER.
+    # (SQLite is dynamically typed and unaffected; this matters on Postgres.)
+    file_size_bytes: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
 
     error_code: Mapped[str | None] = mapped_column(String(64), nullable=True)
     error_message: Mapped[str | None] = mapped_column(Text, nullable=True)
